agent/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
jsonschema>=4.23.0
PyYAML>=6.0.1
feedparser>=6.0.11
python-dotenv>=1.0.1
